+++ CHANGES.md
@@ -6,6 +6,11 @@
 * update FastAPI requirements
 * update minimal python version to 3.6
 
+**breaking change**
+
+* renamed `OptionalHeaders`, `MimeTypes` and `ImageDrivers` enums to the singular form. (https://github.com/developmentseed/titiler/pull/258)
+* renamed `MimeType` to `MediaType` (https://github.com/developmentseed/titiler/pull/258)
+
 ## 0.1.0 (2021-02-17)
 
 * update FastAPI requirements